pyproject.toml
[project]
name = "blokchain"
version = "0.1.0"
description = "Système d'apprentissage collaboratif avec Split Learning et traçabilité Blockchain"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "web3>=6.0.0",
    "py-solc-x>=1.1.1",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
    "pytest>=9.0.2",
]
